# b2b/openapi.yaml
openapi: 3.0.3
info:
  title: NeoMarket B2B (Seller Cabinet)
  version: "1.1.0"
  description: |
    Контракт кабинета продавца: управление товарами, SKU и накладными.
servers:
  - url: https://api.neomarket.local
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/products/{id}:
    parameters:
      - $ref: "#/components/parameters/ProductId"
    get:
      tags: [Products]
      summary: Получить товар со всеми SKU
      responses:
        "200":
          description: Товар найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Products]
      summary: Изменить товар
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: Товар обновлен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProductId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Невалидный запрос
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Сущность не найдена
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: NOT_FOUND
        message:
          type: string
          example: Resource not found
    Category:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    Image:
      type: object
      required: [url, ordering]
      properties:
        url:
          type: string
        ordering:
          type: integer
          minimum: 0
    Characteristic:
      type: object
      required: [name, value]
      properties:
        name:
          type: string
        value:
          type: string
    Sku:
      type: object
      required: [id, product_id, name, price, cost_price, active_quantity, reserved_quantity, images, deleted]
      properties:
        id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        product_title:
          type: string
        name:
          type: string
        price:
          type: integer
          minimum: 0
          description: Цена в копейках
        cost_price:
          type: integer
          minimum: 0
        active_quantity:
          type: integer
          minimum: 0
        reserved_quantity:
          type: integer
          minimum: 0
        images:
          type: array
          items:
            $ref: "#/components/schemas/Image"
        characteristics:
          type: array
          items:
            $ref: "#/components/schemas/Characteristic"
        deleted:
          type: boolean
    Product:
      type: object
      required: [id, title, description, status, category, images, characteristics, skus, skus_count, total_active_quantity, deleted, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [CREATED, ON_MODERATION, MODERATED, BLOCKED, HARD_BLOCKED]
        category:
          $ref: "#/components/schemas/Category"
        images:
          type: array
          items:
            $ref: "#/components/schemas/Image"
        characteristics:
          type: array
          items:
            $ref: "#/components/schemas/Characteristic"
        skus:
          type: array
          items:
            $ref: "#/components/schemas/Sku"
        skus_count:
          type: integer
          minimum: 0
        total_active_quantity:
          type: integer
          minimum: 0
        blocking_reason:
          nullable: true
        field_reports:
          type: array
          items:
            type: object
        deleted:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    UpdateProductRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category_id:
          type: string
          format: uuid
        category_name:
          type: string
        images:
          type: array
          items:
            $ref: "#/components/schemas/Image"
        characteristics:
          type: array
          items:
            $ref: "#/components/schemas/Characteristic"
...