# services/catalog/catalog_api/views.py
import random
import re
from collections import defaultdict
from uuid import UUID

from django.db.models import Q
from django.utils.text import slugify
from drf_spectacular.types import OpenApiTypes
from drf_spectacular.utils import extend_schema, extend_schema_view
from rest_framework import status
from rest_framework.response import Response
from rest_framework.views import APIView

from .models import Category, Product, Sku
from .serializers import (
    CategoryDetailSerializer,
    FacetSerializer,
    CategoryTreeItemSerializer,
    FilterItemSerializer,
    ProductDetailSerializer,
    ProductShortSerializer,
    SkuDetailSerializer,
    SkuShortSerializer,
)

ALLOWED_SORTS = {"rating", "popularity", "price_asc", "price_desc", "date_desc", "discount_desc"}
FILTER_RE = re.compile(r"^filters\[(?P<slug>[^\]]+)\](?:\[(?P<nested>[^\]]+)\])?$")


def _parse_int(value, default, minimum=None, maximum=None):
    try:
        parsed = int(value)
    except (TypeError, ValueError):
        return default

    if minimum is not None:
        parsed = max(minimum, parsed)
    if maximum is not None:
        parsed = min(maximum, parsed)
    return parsed


def _is_uuid(value):
    try:
        UUID(str(value))
        return True
    except (TypeError, ValueError):
        return False


def _error(code, message, http_status):
    return Response({"code": code, "message": message}, status=http_status)


def _normalize_filter_slug(value):
    normalized = slugify(str(value), allow_unicode=True).replace("-", "_")
    return normalized or str(value).strip().lower().replace(" ", "_")


def _parse_filters(query_params):
    filters = {}
    for key, value in query_params.items():
        match = FILTER_RE.match(key)
        if not match:
            continue
        slug = match.group("slug")
        nested = match.group("nested")
        if nested:
            filters.setdefault(slug, {})[nested] = value
        else:
            filters[slug] = value
    return filters


def _base_products_queryset():
    return (
        Product.objects.filter(status=Product.Status.MODERATED)
        .select_related("category", "category__parent")
        .prefetch_related("skus", "images", "attributes")
    )


def _category_or_404(category_id):
    try:
        return Category.objects.select_related("parent").get(id=category_id, is_active=True), None
    except Category.DoesNotExist:
        return None, _error("NOT_FOUND", "Category not found", status.HTTP_404_NOT_FOUND)


def _build_attribute_map(product):
    values = defaultdict(set)
    labels = {}
    for attribute in product.attributes.all():
        slug = _normalize_filter_slug(attribute.name)
        labels.setdefault(slug, str(attribute.name))
        values[slug].add(str(attribute.value))

    for sku in product.skus.all():
        for name, value in (sku.attributes or {}).items():
            slug = _normalize_filter_slug(name)
            labels.setdefault(slug, str(name))
            values[slug].add(str(value))

    return values, labels


def _product_matches_filters(product, filters):
    attribute_values, _labels = _build_attribute_map(product)
    sku_prices = [int(sku.price) for sku in product.skus.all()]
    max_stock = max((int(sku.active_quantity or 0) for sku in product.skus.all()), default=0)

    for slug, raw_value in filters.items():
        if slug == "availability":
            if str(raw_value).lower() in {"1", "true", "yes", "on"} and max_stock <= 0:
                return False
            continue

        if slug == "price" and isinstance(raw_value, dict):
            min_price = _parse_int(raw_value.get("min"), default=None)
            max_price = _parse_int(raw_value.get("max"), default=None)
            if min_price is not None and sku_prices and max(sku_prices) < min_price:
                return False
            if max_price is not None and sku_prices and min(sku_prices) > max_price:
                return False
            continue

        requested_values = raw_value if isinstance(raw_value, list) else str(raw_value).split(",")
        requested_values = [str(item).strip() for item in requested_values if str(item).strip()]
        if not requested_values:
            continue
        available_values = {str(item) for item in attribute_values.get(slug, set())}
        if not available_values.intersection(requested_values):
            return False

    return True


def _sorted_products(products, sort):
    if sort == "price_asc":
        return sorted(products, key=lambda item: (min((sku.price for sku in item.skus.all()), default=0), -item.created_at.timestamp()))
    if sort == "price_desc":
        return sorted(products, key=lambda item: (max((sku.price for sku in item.skus.all()), default=0), item.created_at.timestamp()), reverse=True)
    if sort in {"rating", "popularity", "discount_desc", "date_desc"}:
        return sorted(products, key=lambda item: item.created_at, reverse=True)
    return sorted(products, key=lambda item: item.created_at, reverse=True)


def _category_chain(category):
    chain = []
    seen = set()
    current = category
    while current:
        if current.id in seen:
            return None, _error("orphan_node", "category hierarchy is broken", status.HTTP_422_UNPROCESSABLE_ENTITY)
        seen.add(current.id)
        chain.append(current)
        current = current.parent
    chain.reverse()
    return chain, None


def _category_url(chain, current):
    slug_path = "/".join(item.slug for item in chain[: current + 1])
    return f"/catalog/{slug_path}"


@extend_schema_view(
    get=extend_schema(operation_id="catalog_list_products", responses=OpenApiTypes.OBJECT),
)
class ProductListView(APIView):
    def get(self, request):
        limit = _parse_int(request.query_params.get("limit", 20), default=20, minimum=1, maximum=100)
        offset = _parse_int(request.query_params.get("offset", 0), default=0, minimum=0)
        sort = request.query_params.get("sort", "rating")
        if sort not in ALLOWED_SORTS:
            message = "Invalid sort parameter. Allowed: rating, popularity, price_asc, price_desc, date_desc, discount_desc"
            return _error("INVALID_REQUEST", message, status.HTTP_400_BAD_REQUEST)

        queryset = _base_products_queryset()
        ids_param = request.query_params.get("ids")
        sku_ids_param = request.query_params.get("sku_ids") or request.query_params.get("sku_id")
        if ids_param:
            raw_ids = [item.strip() for item in ids_param.split(",") if item.strip()]
            parsed_ids = []
            for raw_id in raw_ids:
                if not _is_uuid(raw_id):
                    return _error("INVALID_REQUEST", "ids must contain UUID values", status.HTTP_400_BAD_REQUEST)
                parsed_ids.append(raw_id)
            items = list(queryset.filter(id__in=parsed_ids))
            items_by_id = {str(item.id): item for item in items}
            ordered_items = [items_by_id[item_id] for item_id in raw_ids if item_id in items_by_id]
            serializer = ProductDetailSerializer(ordered_items, many=True)
            return Response({"items": serializer.data, "total_count": len(ordered_items), "limit": len(ordered_items), "offset": 0})
        if sku_ids_param:
            raw_sku_ids = [item.strip() for item in sku_ids_param.split(",") if item.strip()]
            parsed_sku_ids = []
            for raw_id in raw_sku_ids:
                if not _is_uuid(raw_id):
                    return _error("INVALID_REQUEST", "sku_ids must contain UUID values", status.HTTP_400_BAD_REQUEST)
                parsed_sku_ids.append(raw_id)
            items = list(queryset.filter(skus__id__in=parsed_sku_ids).distinct())
            serializer = ProductDetailSerializer(items, many=True)
            return Response({"items": serializer.data, "total_count": len(items), "limit": len(items), "offset": 0})

        category_id = request.query_params.get("category_id")
        if category_id:
            if not _is_uuid(category_id):
                return _error("INVALID_REQUEST", "Invalid category_id", status.HTTP_400_BAD_REQUEST)
            _category, error = _category_or_404(category_id)
            if error:
                return error
            queryset = queryset.filter(category_id=category_id)

        search = request.query_params.get("search")
        if search:
            search = search.strip()
            if len(search) < 3:
                return _error("INVALID_REQUEST", "Search query must be at least 3 characters", status.HTTP_400_BAD_REQUEST)
            if len(search) > 255:
                return _error("INVALID_REQUEST", "Search query must be at most 255 characters", status.HTTP_400_BAD_REQUEST)
            queryset = queryset.filter(Q(title__icontains=search) | Q(description__icontains=search))

        filters = _parse_filters(request.query_params)
        products = [product for product in queryset if _product_matches_filters(product, filters)]
        products = _sorted_products(products, sort)

        total_count = len(products)
        items = products[offset : offset + limit]

        serializer = ProductShortSerializer(items, many=True)
        return Response(
            {
                "total_count": total_count,
                "limit": limit,
                "offset": offset,
                "items": serializer.data,
            }
        )


@extend_schema_view(
    get=extend_schema(operation_id="catalog_get_product", responses=ProductDetailSerializer),
)
class ProductDetailView(APIView):
    def get(self, request, id):
        try:
            product = _base_products_queryset().get(
                id=id,
                status=Product.Status.MODERATED,
            )
        except Product.DoesNotExist:
            return _error("NOT_FOUND", "Product not found", status.HTTP_404_NOT_FOUND)

        serializer = ProductDetailSerializer(product)
        return Response(serializer.data)


@extend_schema_view(
    get=extend_schema(operation_id="catalog_list_similar_products", responses=OpenApiTypes.OBJECT),
)
class ProductSimilarView(APIView):
    def get(self, request, id):
        limit = _parse_int(request.query_params.get("limit", 8), default=8, minimum=1, maximum=20)
        offset = _parse_int(request.query_params.get("offset", 0), default=0, minimum=0)

        try:
            product = _base_products_queryset().get(id=id, status=Product.Status.MODERATED)
        except Product.DoesNotExist:
            return _error("NOT_FOUND", "Product not found", status.HTTP_404_NOT_FOUND)

        category_id = request.query_params.get("category") or str(product.category_id)
        if not _is_uuid(category_id):
            return _error("INVALID_REQUEST", "Nonexistent category id", status.HTTP_400_BAD_REQUEST)

        category_products = [item for item in _base_products_queryset().filter(category_id=category_id) if item.id != product.id]
        if len(category_products) < limit and product.category.parent_id:
            parent_products = [
                item
                for item in _base_products_queryset().filter(category=product.category.parent)
                if item.id != product.id and item.id not in {candidate.id for candidate in category_products}
            ]
            category_products.extend(parent_products)

        randomizer = random.Random(f"{product.id}:{category_id}")
        randomizer.shuffle(category_products)
        total_count = len(category_products)
        items = category_products[offset : offset + limit]
        serializer = ProductShortSerializer(items, many=True)

        return Response(
            {
                "items": serializer.data,
                "total_count": total_count,
                "limit": limit,
                "offset": offset,
            }
        )


@extend_schema_view(
    get=extend_schema(operation_id="catalog_list_product_skus", responses=SkuShortSerializer(many=True)),
)
class ProductSkuListView(APIView):
    def get(self, request, product_id):
        skus = Sku.objects.filter(product_id=product_id, product__status=Product.Status.MODERATED).select_related("product")
        serializer = SkuShortSerializer(skus, many=True)
        return Response(serializer.data)


@extend_schema_view(
    get=extend_schema(operation_id="catalog_get_product_sku", responses=SkuDetailSerializer),
)
class ProductSkuDetailView(APIView):
    def get(self, request, product_id, sku_id):
        try:
            sku = Sku.objects.select_related("product").get(id=sku_id, product_id=product_id, product__status=Product.Status.MODERATED)
        except Sku.DoesNotExist:
            return _error("NOT_FOUND", "SKU not found", status.HTTP_404_NOT_FOUND)

        serializer = SkuDetailSerializer(sku)
        return Response(serializer.data)


@extend_schema_view(
    get=extend_schema(operation_id="catalog_get_categories_tree", responses=OpenApiTypes.OBJECT),
)
class CategoryTreeView(APIView):
    def get(self, request):
        roots = Category.objects.filter(parent__isnull=True, is_active=True).order_by("name")
        serializer = CategoryTreeItemSerializer(roots, many=True)
        return Response({"items": serializer.data})


@extend_schema_view(
    get=extend_schema(operation_id="catalog_get_category", responses=CategoryDetailSerializer),
)
class CategoryDetailView(APIView):
    def get(self, request, id):
        include_product_count = str(request.query_params.get("include_product_count", "false")).lower() == "true"
        category, error = _category_or_404(id)
        if error:
            return error

        serializer = CategoryDetailSerializer(category, context={"include_product_count": include_product_count})
        return Response(serializer.data)


@extend_schema_view(
    get=extend_schema(operation_id="catalog_get_category_filters", responses=OpenApiTypes.OBJECT),
)
class CategoryFiltersView(APIView):
    def get(self, request, id):
        category, error = _category_or_404(id)
        if error:
            return error

        products = list(_base_products_queryset().filter(category=category))
        skus = [sku for product in products for sku in product.skus.all()]
        min_price = min((int(sku.price) for sku in skus), default=0)
        max_price = max((int(sku.price) for sku in skus), default=0)

        dynamic_values = defaultdict(set)
        dynamic_labels = {}
        for product in products:
            attribute_values, labels = _build_attribute_map(product)
            for slug, values in attribute_values.items():
                dynamic_labels.setdefault(slug, labels.get(slug, slug))
                dynamic_values[slug].update(values)

        items = [
            {"slug": "availability", "name": "В наличии", "type": "switch"},
            {"slug": "price", "name": "Цена", "type": "range", "min": min_price, "max": max_price},
        ]
        for slug in sorted(dynamic_values):
            values = sorted(dynamic_values[slug])
            if values:
                items.append({"slug": slug, "name": dynamic_labels.get(slug, slug), "type": "list", "value": values})

        serializer = FilterItemSerializer(items, many=True)
        return Response({"items": serializer.data})


@extend_schema_view(
    get=extend_schema(operation_id="catalog_get_facets", responses=OpenApiTypes.OBJECT),
)
class CatalogFacetsView(APIView):
    def get(self, request):
        category_id = request.query_params.get("category_id")
        if not category_id or not _is_uuid(category_id):
            return _error("INVALID_REQUEST", "category_id must be provided", status.HTTP_400_BAD_REQUEST)
        category, error = _category_or_404(category_id)
        if error:
            return error

        filters = _parse_filters(request.query_params)
        products = [product for product in _base_products_queryset().filter(category=category) if _product_matches_filters(product, filters)]

        counters = defaultdict(lambda: defaultdict(int))
        for product in products:
            attribute_values, _labels = _build_attribute_map(product)
            for slug, values in attribute_values.items():
                for value in values:
                    counters[slug][value] += 1

        facets = [{"name": slug, "values": [{"value": value, "count": count} for value, count in sorted(values.items())]} for slug, values in sorted(counters.items())]
        serializer = FacetSerializer(facets, many=True)
        return Response({"category_id": str(category.id), "facets": serializer.data})


@extend_schema_view(
    get=extend_schema(operation_id="catalog_get_breadcrumbs", responses=OpenApiTypes.OBJECT),
)
class BreadcrumbsView(APIView):
    def get(self, request):
        category_id = request.query_params.get("category_id")
        product_id = request.query_params.get("product_id")
        if category_id and product_id:
            return _error("ambiguous_param", "only one of category_id or product_id must be provided", status.HTTP_400_BAD_REQUEST)
        if not category_id and not product_id:
            return _error("missing_param", "category_id or product_id must be provided", status.HTTP_400_BAD_REQUEST)

        if category_id:
            if not _is_uuid(category_id):
                return _error("INVALID_REQUEST", "Invalid category_id", status.HTTP_400_BAD_REQUEST)
            category, error = _category_or_404(category_id)
            if error:
                return error
            chain, error = _category_chain(category)
            if error:
                return error
            data = [
                {
                    "id": str(item.id),
                    "slug": item.slug,
                    "name": item.name,
                    "url": _category_url(chain, index),
                    "level": index,
                    "is_current": index == len(chain) - 1,
                }
                for index, item in enumerate(chain)
            ]
            return Response({"data": data, "meta": {"resolved_via": "category_id", "category_id": str(category.id)}})

        if not _is_uuid(product_id):
            return _error("INVALID_REQUEST", "Invalid product_id", status.HTTP_400_BAD_REQUEST)
        try:
            product = _base_products_queryset().get(id=product_id, status=Product.Status.MODERATED)
        except Product.DoesNotExist:
            return _error("NOT_FOUND", "Product not found", status.HTTP_404_NOT_FOUND)
        chain, error = _category_chain(product.category)
        if error:
            return error
        data = [
            {
                "id": str(item.id),
                "slug": item.slug,
                "name": item.name,
                "url": _category_url(chain, index),
                "level": index,
                "is_current": False,
            }
            for index, item in enumerate(chain)
        ]
        data.append(
            {
                "id": str(product.id),
                "slug": slugify(product.title, allow_unicode=True),
                "name": product.title,
                "url": f"/products/{product.id}",
                "level": len(chain),
                "is_current": True,
            }
        )
        return Response({"data": data, "meta": {"resolved_via": "product_id", "product_id": str(product.id), "category_id": str(product.category_id)}})